refactor: create apis/parameters.ts and move getParameter() there both for code and tests
