+++ src/api.ts
@@ -1,8 +1,8 @@
 import * as vscode from 'vscode';
-//import {isEmpty, get} from 'lodash';
 import { performAction } from './actions/client';
 import { ActionsController } from './actions/controller';
 import { ExecuteAction, SnippetAction, CommandAction, FileAction } from './actions/impl';
+import { getParameter } from './apis/parameters';
 import { getLogger } from './logger/logger';
 export * from "./actions/interfaces";
 
@@ -38,30 +38,7 @@ export const bas = {
     },
 
     async getParameter(parameterName: string) : Promise<string | undefined> {
-        // const onBasEnv = "env.WS_BASE_URL";
-        // const isInBAS = !isEmpty(get(process, onBasEnv));
-        //const isInBAS = get(process, onBasEnv)?.is;
-        // if (!isInBAS) {
-        //     getLogger().trace("Running in VS Code, so returning undefined.");
-        //     return undefined;
-        // }
-        // getLogger().trace("Running on BAS.");
-
-        const optionalRequire = require("optional-require")(require);
-        const noSapPlugin = "noSapPlugin";
-        const sapPlugin = optionalRequire('@sap/plugin') || noSapPlugin;
-        if (sapPlugin === noSapPlugin) {
-            getLogger().error("Failed to load @sap/plugin, so returning undefined.");
-            return undefined;
-        } 
-        getLogger().trace("@sap/plugin successfully loaded.");
-
-        const configuration = await sapPlugin.window.configuration();
-        getLogger().trace("Configuration successfully received.", {configuration});
-
-        const parameterValue = configuration?.[parameterName];
-        getLogger().trace(`configuration[${parameterName}]=${parameterValue}`, {parameterName, parameterValue});
-        return parameterValue;
+        return getParameter(parameterName);
     },
 
     actions: {
